Изменение конструктора BigNumber с параметром long long

+++ BigNumber.h
@@ -7,7 +7,7 @@ class BigNumber
 public:
 	BigNumber();
 	BigNumber(int32_t num);
-	BigNumber(long long num);
+	BigNumber(int64_t num);
 	BigNumber operator~();
 	BigNumber& operator^=(const BigNumber& big_number);
 	BigNumber operator^(const BigNumber& big_number);
